AG-12912 - Create start end scene for textualStartEndScene and shapes with start/end points
