+++ packages/ag-charts-enterprise/src/features/annotations/scenes/startEndScene.ts
@@ -1,0 +1,147 @@
+import { type AgAnnotationHandleStyles, _ModuleSupport, _Scene, _Util } from 'ag-charts-community';
+
+import type { AnnotationContext, Coords, LineCoords } from '../annotationTypes';
+import type { StartEndProperties } from '../properties/startEndProperties';
+import { validateDatumPoint } from '../utils/validation';
+import { convertLine, invertCoords } from '../utils/values';
+import { DivariantHandle } from './handle';
+import { LinearScene } from './linearScene';
+
+const { Vec2 } = _Util;
+
+export abstract class StartEndScene<Datum extends StartEndProperties> extends LinearScene<Datum> {
+    override activeHandle?: 'start' | 'end';
+
+    protected readonly start = new DivariantHandle();
+    protected readonly end = new DivariantHandle();
+
+    protected anchor: _ModuleSupport.ToolbarAnchor = {
+        x: 0,
+        y: 0,
+        position: 'above',
+    };
+
+    public update(datum: Datum, context: AnnotationContext) {
+        const coords = convertLine(datum, context);
+
+        if (coords == null) {
+            return;
+        }
+
+        this.updateHandles(datum, coords);
+
+        this.anchor = this.updateAnchor(datum, coords, context);
+    }
+
+    override toggleHandles(show: boolean | Partial<Record<'start' | 'end', boolean>>) {
+        if (typeof show === 'boolean') {
+            show = { start: show, end: show };
+        }
+
+        this.start.visible = show.start ?? true;
+        this.end.visible = show.end ?? true;
+
+        this.start.toggleHovered(this.activeHandle === 'start');
+        this.end.toggleHovered(this.activeHandle === 'end');
+    }
+
+    override toggleActive(active: boolean) {
+        this.toggleHandles(active);
+        this.start.toggleActive(active);
+        this.end.toggleActive(active);
+    }
+
+    override dragHandle(datum: Datum, target: Coords, context: AnnotationContext) {
+        const { activeHandle, dragState } = this;
+
+        if (!activeHandle || !dragState) return;
+
+        this[activeHandle].toggleDragging(true);
+        const coords = Vec2.add(dragState.end, Vec2.sub(target, dragState.offset));
+        const point = invertCoords(coords, context);
+
+        if (!validateDatumPoint(context, point)) return;
+
+        datum[activeHandle].x = point.x;
+        datum[activeHandle].y = point.y;
+    }
+
+    override stopDragging() {
+        this.start.toggleDragging(false);
+        this.end.toggleDragging(false);
+    }
+
+    override getAnchor() {
+        return this.anchor;
+    }
+
+    override getCursor() {
+        if (this.activeHandle == null) return 'pointer';
+    }
+
+    override containsPoint(x: number, y: number) {
+        const { start, end } = this;
+
+        this.activeHandle = undefined;
+
+        if (start.containsPoint(x, y)) {
+            this.activeHandle = 'start';
+            return true;
+        }
+
+        if (end.containsPoint(x, y)) {
+            this.activeHandle = 'end';
+            return true;
+        }
+
+        return false;
+    }
+
+    protected updateHandles(datum: Datum, coords: LineCoords, bbox?: _Scene.BBox) {
+        this.start.update({
+            ...this.getHandleStyles(datum, 'start'),
+            ...this.getHandleCoords(datum, coords, 'start'),
+        });
+        this.end.update({
+            ...this.getHandleStyles(datum, 'end'),
+            ...this.getHandleCoords(datum, coords, 'end', bbox),
+        });
+
+        this.start.toggleLocked(datum.locked ?? false);
+        this.end.toggleLocked(datum.locked ?? false);
+    }
+
+    protected updateAnchor(
+        _datum: Datum,
+        coords: LineCoords,
+        context: AnnotationContext,
+        _bbox?: _Scene.BBox
+    ): _ModuleSupport.ToolbarAnchor {
+        return {
+            x: coords.x1 + context.seriesRect.x,
+            y: coords.y1 + context.seriesRect.y,
+            position: this.anchor.position,
+        };
+    }
+
+    protected getHandleCoords(
+        _datum: Datum,
+        coords: LineCoords,
+        handle: 'start' | 'end',
+        _bbox?: _Scene.BBox
+    ): _Util.Vec2 {
+        return {
+            x: handle === 'start' ? coords.x1 : coords.x2,
+            y: handle === 'start' ? coords.y1 : coords.y2,
+        };
+    }
+
+    protected getHandleStyles(datum: Datum, _handle: 'start' | 'end'): AgAnnotationHandleStyles {
+        return {
+            fill: datum.handle.fill,
+            stroke: datum.handle.stroke,
+            strokeOpacity: datum.handle.strokeOpacity,
+            strokeWidth: datum.handle.strokeWidth,
+        };
+    }
+}
