+++ packages/ag-charts-enterprise/src/features/annotations/scenes/textualStartEndScene.ts
@@ -1,23 +1,17 @@
 import { type AgAnnotationHandleStyles, _ModuleSupport, _Scene, _Util } from 'ag-charts-community';
 
-import type { AnnotationContext, Coords, LineCoords } from '../annotationTypes';
+import type { AnnotationContext, LineCoords } from '../annotationTypes';
 import type { TextualStartEndProperties } from '../properties/textualStartEndProperties';
-import { DivariantHandle } from '../scenes/handle';
-import { LinearScene } from '../scenes/linearScene';
+import { StartEndScene } from '../scenes/startEndScene';
 import { getBBox, updateTextNode, wrapText } from '../text/util';
-import { validateDatumPoint } from '../utils/validation';
-import { convertLine, invertCoords } from '../utils/values';
+import { convertLine } from '../utils/values';
 
-const { Vec2 } = _Util;
-
-export abstract class TextualStartEndScene<Datum extends TextualStartEndProperties> extends LinearScene<Datum> {
+export abstract class TextualStartEndScene<Datum extends TextualStartEndProperties> extends StartEndScene<Datum> {
     override activeHandle?: 'start' | 'end';
 
     protected readonly label = new _Scene.Text({ zIndex: 1 });
-    protected readonly start = new DivariantHandle();
-    protected readonly end = new DivariantHandle();
 
-    protected anchor: _ModuleSupport.ToolbarAnchor = {
+    protected override anchor: _ModuleSupport.ToolbarAnchor = {
         x: 0,
         y: 0,
         position: 'above-left',
@@ -29,7 +23,7 @@ export abstract class TextualStartEndScene<Datum extends TextualStartEndProperti
         this.markDirty(this, _Scene.RedrawType.MINOR);
     }
 
-    public update(datum: Datum, context: AnnotationContext) {
+    public override update(datum: Datum, context: AnnotationContext) {
         const coords = convertLine(datum, context);
 
         if (coords == null) {
@@ -39,74 +33,16 @@ export abstract class TextualStartEndScene<Datum extends TextualStartEndProperti
         const bbox = this.getTextBBox(datum, coords);
 
         this.updateLabel(datum, bbox, coords);
-        this.updateHandles(datum, bbox, coords);
+        this.updateHandles(datum, coords, bbox);
         this.updateShape(datum, bbox, coords);
 
-        this.anchor = this.updateAnchor(datum, bbox, context);
-    }
-
-    override toggleHandles(show: boolean | Partial<Record<'start' | 'end', boolean>>) {
-        if (typeof show === 'boolean') {
-            show = { start: show, end: show };
-        }
-
-        this.start.visible = show.start ?? true;
-        this.end.visible = show.end ?? true;
-
-        this.start.toggleHovered(this.activeHandle === 'start');
-        this.end.toggleHovered(this.activeHandle === 'end');
-    }
-
-    override toggleActive(active: boolean) {
-        this.toggleHandles(active);
-        this.start.toggleActive(active);
-        this.end.toggleActive(active);
-    }
-
-    override dragHandle(datum: Datum, target: Coords, context: AnnotationContext) {
-        const { activeHandle, dragState } = this;
-
-        if (!activeHandle || !dragState) return;
-
-        this[activeHandle].toggleDragging(true);
-        const coords = Vec2.add(dragState.end, Vec2.sub(target, dragState.offset));
-        const point = invertCoords(coords, context);
-
-        if (!validateDatumPoint(context, point)) return;
-
-        datum[activeHandle].x = point.x;
-        datum[activeHandle].y = point.y;
-    }
-
-    override stopDragging() {
-        this.start.toggleDragging(false);
-        this.end.toggleDragging(false);
-    }
-
-    override getAnchor() {
-        return this.anchor;
-    }
-
-    override getCursor() {
-        if (this.activeHandle == null) return 'pointer';
+        this.anchor = this.updateAnchor(datum, coords, context, bbox);
     }
 
     override containsPoint(x: number, y: number) {
-        const { start, end, label } = this;
+        const { label } = this;
 
-        this.activeHandle = undefined;
-
-        if (start.containsPoint(x, y)) {
-            this.activeHandle = 'start';
-            return true;
-        }
-
-        if (end.containsPoint(x, y)) {
-            this.activeHandle = 'end';
-            return true;
-        }
-
-        return label.containsPoint(x, y);
+        return super.containsPoint(x, y) || label.containsPoint(x, y);
     }
 
     protected getTextBBox(datum: Datum, coords: LineCoords) {
@@ -126,30 +62,8 @@ export abstract class TextualStartEndScene<Datum extends TextualStartEndProperti
         updateTextNode(this.label, wrappedText, isPlaceholder, datum, this.getLabelCoords(datum, bbox, coords));
     }
 
-    protected updateHandles(datum: Datum, bbox: _Scene.BBox, coords: LineCoords) {
-        this.start.update({
-            ...this.getHandleStyles(datum, 'start'),
-            ...this.getHandleCoords(datum, bbox, coords, 'start'),
-        });
-        this.end.update({
-            ...this.getHandleStyles(datum, 'end'),
-            ...this.getHandleCoords(datum, bbox, coords, 'end'),
-        });
-
-        this.start.toggleLocked(datum.locked ?? false);
-        this.end.toggleLocked(datum.locked ?? false);
-    }
-
     protected updateShape(_datum: Datum, _textBBox: _Scene.BBox, _coords: LineCoords) {
         // Shapes should be implemented by the extending annotation type class
-    }
-
-    protected updateAnchor(_datum: Datum, bbox: _Scene.BBox, context: AnnotationContext): _ModuleSupport.ToolbarAnchor {
-        return {
-            x: bbox.x + context.seriesRect.x,
-            y: bbox.y + context.seriesRect.y - bbox.height,
-            position: this.anchor.position,
-        };
     }
 
     protected getLabelCoords(_datum: Datum, _bbox: _Scene.BBox, coords: LineCoords): _Util.Vec2 {
@@ -159,24 +73,11 @@ export abstract class TextualStartEndScene<Datum extends TextualStartEndProperti
         };
     }
 
-    protected getHandleCoords(
-        _datum: Datum,
-        _bbox: _Scene.BBox,
-        coords: LineCoords,
-        handle: 'start' | 'end'
-    ): _Util.Vec2 {
+    protected override getHandleStyles(datum: Datum, handle: 'start' | 'end'): AgAnnotationHandleStyles {
         return {
-            x: handle === 'start' ? coords.x1 : coords.x2,
-            y: handle === 'start' ? coords.y1 : coords.y2,
-        };
-    }
+            ...super.getHandleStyles(datum, handle),
 
-    protected getHandleStyles(datum: Datum, _handle: 'start' | 'end'): AgAnnotationHandleStyles {
-        return {
-            fill: datum.handle.fill,
             stroke: datum.handle.stroke ?? datum.color,
-            strokeOpacity: datum.handle.strokeOpacity,
-            strokeWidth: datum.handle.strokeWidth,
         };
     }
 }
